* develop(1105-sftp-support): adding SFTP support

* Merge branch 'main' into 1105-sftp-support

* develop(1105-sftp-support): fix SSM parameter import/export

* develop(1105-sftp-support): add SFTP server

* develop(1105-sftp-support): add SFTP server

* develop(1105-sftp-support): add SFTP server

* develop(1105-sftp-support): add SFTP server

* develop(1105-sftp-support): add SFTP server

* develop(1105-sftp-support): add SFTP server

* Merge branch 'main' into 1105-sftp-support

* develop(1105-sftp-support): add SFTP server

* develop(1105-sftp-support): add Cognito identity provider details

* Merge branch 'main' into 1105-sftp-support

* develop(1105-sftp-support): add Cognito identity provider details

* Merge branch 'main' into 1105-sftp-support

* develop(1105-sftp-support): add Cognito identity provider details

* develop(1105-sftp-support): change logging

* develop(1105-sftp-support): change logging

* develop(1105-sftp-support): change logging

* Merge remote-tracking branch 'refs/remotes/origin/main' into 1105-sft…

* develop(1105-sftp-support): change logging

* develop(1105-sftp-support): change logging

* develop(1105-sftp-support): change logging

* Merge branch 'main' into 1105-sftp-support

* Merge branch 'main' into 1105-sftp-support

* develop(1105-sftp-support): change logging

* develop(1105-sftp-support): change logging

* develop(1105-sftp-support): change logging

* develop(1105-sftp-support): change logging

* develop(1105-sftp-support): change logging

* develop(1105-sftp-support): fix SQS message attributes

* develop(1105-sftp-support): fix SQS message attributes

* develop(1105-sftp-support): fix SQS message attributes

* develop(1105-sftp-support): fix SQS message attributes

* Merge branch 'main' into 1105-sftp-support

* develop(1105-sftp-support): add FTP users protocols

* develop(1105-sftp-support): fix queue/topic import

* develop(1105-sftp-support): fix queue/topic import

* develop(1105-sftp-support): fix queue/topic import

* Merge branch 'main' into 1105-sftp-support

* develop(1105-sftp-support): fix queue/topic import

* develop(1105-sftp-support): fix queue/topic import

* develop(1105-sftp-support): fix queue/topic import

* develop(1105-sftp-support): fix awslocal

* develop(1105-sftp-support): new import request

* Merge branch 'main' into 1105-sftp-support

* develop(1105-sftp-support): new import request

* develop(1105-sftp-support): new import request

* develop(1105-sftp-support): new import request

* develop(1105-sftp-support): new import request

* Merge branch 'main' into 1105-sftp-support

* Merge branch 'main' into 1105-sftp-support

* develop(1105-sftp-support): new import request

* Merge branch 'main' into 1105-sftp-support

* Merge branch 'main' into 1105-sftp-support

* develop(1105-sftp-support): new import request

* develop(1105-sftp-support): new import request

* develop(1105-sftp-support): new import request

* develop(1105-sftp-support): new import request

* develop(1105-sftp-support): new import request

* Merge branch 'main' into 1105-sftp-support

* develop(1105-sftp-support): new import request

* develop(1105-sftp-support): new import request

* develop(1105-sftp-support): new import request

* develop(1105-sftp-support): new import request

* develop(1105-sftp-support): new import request

* Merge branch 'main' into 1105-sftp-support

* Merge remote-tracking branch 'origin/1105-sftp-support' into 1105-sft…

+++ src/dto/include/awsmock/dto/sqs/internal/ListMessageCountersRequest.h
@@ -28,12 +28,12 @@ namespace AwsMock::Dto::SQS {
         /**
          * Page size
          */
-        int pageSize = 10;
+        long pageSize = 10;
 
         /**
          * Page index
          */
-        int pageIndex = 0;
+        long pageIndex = 0;
 
         /**
          * Sort column
@@ -46,8 +46,8 @@ namespace AwsMock::Dto::SQS {
             ListMessageCountersRequest r;
             r.queueArn = v.at("queueArn").as_string();
             r.prefix = v.at("prefix").as_string();
-            r.pageSize = static_cast<int>(v.at("pageSize").as_int64());
-            r.pageIndex = static_cast<int>(v.at("pageIndex").as_int64());
+            r.pageSize = v.at("pageSize").as_int64();
+            r.pageIndex = v.at("pageIndex").as_int64();
             r.sortColumns = boost::json::value_to<std::vector<Common::SortColumn>>(v.at("sortColumns"));
             return r;
         }
